TMA/pyproject.toml
[tool.poetry]
name = "tma_asset_generator"
version = "0.1.0"
description = "A package to generate prompts and images for Telegram using OpenAI's APIs."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "tma_asset_generator" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
openai = "^0.27.0"
aiohttp = "^3.8.1"
pandas = "^1.4.4"
Pillow = "^9.2.0"
nest-asyncio = "^1.5.6"
duckdb = "^1.0.0"
click = "^8.1.3"
numpy = "^1.25.0"

[tool.poetry.scripts]
tma-asset-generator = "cli:main"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=65.5.0", "wheel", "cmake>=3.21.0", "protobuf>=3.0.0"]
build-backend = "poetry.core.masonry.api"
